Convert digi sounds manually instead of using SDL's audio converter
This fixes some crackling and popping noises, apparently occurring due to bugs in SDL's built-in audio conversion routines.

+++ src/types.h
@@ -561,7 +561,7 @@ typedef struct ogg_type {
 
 typedef struct converted_audio_type {
 	int length;
-	byte samples[0];
+	short samples[0];
 } converted_audio_type;
 
 typedef struct sound_buffer_type {
